+++ samples/Demo/Beef.Demo.Database/Migrations/20210428-232715-create-demo-eventoutbox.sql
@@ -4,7 +4,7 @@ CREATE TABLE [Demo].[EventOutbox] (
    */
 
   [EventOutboxId] BIGINT IDENTITY (1, 1) NOT NULL PRIMARY KEY NONCLUSTERED ([EventOutboxId] ASC),
-  [EnqueuedDate] DATETIME NOT NULL,
-  [DequeuedDate] DATETIME NULL,
+  [EnqueuedDate] DATETIME2 NOT NULL,
+  [DequeuedDate] DATETIME2 NULL,
   CONSTRAINT [IX_Demo__DequeuedDate] UNIQUE CLUSTERED ([DequeuedDate], [EventOutboxId])
 );