+++ samples/Demo/Beef.Demo.Test/ContactTest.cs
@@ -2,6 +2,8 @@
 using Beef.Demo.Common.Agents;
 using Beef.Demo.Common.Entities;
 using Beef.Test.NUnit;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
 using NUnit.Framework;
 using System.Net;
 
@@ -39,7 +41,6 @@ namespace Beef.Demo.Test
         public void A120_Update()
         {
             using var agentTester = AgentTester.CreateWaf<Startup>();
-
             var r = agentTester.Test<ContactAgent, Contact>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .ExpectValue((t) => new Contact { Id = 1.ToGuid(), FirstName = "Jenny", LastName = "Cuthbert" })
@@ -83,6 +84,7 @@ namespace Beef.Demo.Test
             var v = r.Value.Result[0];
             v.LastName += "X";
 
+            // Update and ensure that the etag has changed as a result.
             var r2 = agentTester.Test<ContactAgent, Contact>()
                 .ExpectStatusCode(HttpStatusCode.OK)
                 .ExpectValue((t) => v)
@@ -90,6 +92,35 @@ namespace Beef.Demo.Test
 
             Assert.NotNull(r2.Response.Headers?.ETag?.Tag);
             Assert.AreNotEqual(etag, r2.Response.Headers?.ETag?.Tag);
+        }
+
+        [Test, TestSetUp]
+        public void A200_RaiseEvent_Failure()
+        {
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
+            var r = agentTester.Test<ContactAgent>()
+                .ExpectStatusCode(HttpStatusCode.InternalServerError)
+                .Run(a => a.RaiseEventAsync(true));
+
+            var db = new Beef.Demo.Business.Data.Database(agentTester.WebApplicationFactory.Services.GetService<IConfiguration>()["ConnectionStrings:BeefDemo"]);
+            var count = db.SqlStatement("SELECT COUNT(*) FROM [Demo].[EventOutboxData] WHERE [Subject] = 'Contact' and [Action] = 'Made'").ScalarAsync<int>().GetAwaiter().GetResult();
+            Assert.AreEqual(0, count);
+        }
+
+        [Test, TestSetUp]
+        public void A210_RaiseEvent_Success()
+        {
+            using var agentTester = AgentTester.CreateWaf<Startup>();
+
+            var r = agentTester.Test<ContactAgent>()
+                .ExpectStatusCode(HttpStatusCode.NoContent)
+                .ExpectEvent("Contact", "Made")
+                .Run(a => a.RaiseEventAsync(false));
+
+            var db = new Beef.Demo.Business.Data.Database(agentTester.WebApplicationFactory.Services.GetService<IConfiguration>()["ConnectionStrings:BeefDemo"]);
+            var count = db.SqlStatement("SELECT COUNT(*) FROM [Demo].[EventOutboxData] WHERE [Subject] = 'Contact' and [Action] = 'Made'").ScalarAsync<int>().GetAwaiter().GetResult();
+            Assert.AreEqual(1, count);
         }
     }
 }