+++ tools/Beef.CodeGen.Core/Config/Entity/OperationConfig.cs
@@ -165,6 +165,14 @@ operations: [
         public bool? DataExtensions { get; set; }
 
         /// <summary>
+        /// Indicates whether a `System.TransactionScope` should be created and orchestrated at the `Data`-layer.
+        /// </summary>
+        [JsonProperty("dataTransaction", DefaultValueHandling = DefaultValueHandling.Ignore)]
+        [PropertySchema("Data", Title = "Indicates whether a `System.TransactionScope` should be created and orchestrated at the `Data`-layer.",
+            Description = "Where using an `EventOutbox` this is ignored as it is implied through its usage.")]
+        public bool? DataTransaction { get; set; }
+
+        /// <summary>
         /// Gets or sets the database stored procedure name.
         /// </summary>
         [JsonProperty("databaseStoredProc", DefaultValueHandling = DefaultValueHandling.Ignore)]
@@ -283,6 +291,14 @@ operations: [
         [PropertySchema("Events", Title = "The layer to add logic to publish an event for a `Create`, `Update` or `Delete` operation.", IsImportant = true, Options = new string[] { "None", "DataSvc", "Data" },
             Description = "Defaults to the `Entity.EventPublish` configuration property (inherits) where not specified. Used to enable the sending of messages to the likes of EventGrid, Service Broker, SignalR, etc.")]
         public string? EventPublish { get; set; }
+
+        /// <summary>
+        /// Gets or sets the data-tier event outbox persistence technology (where the events will be transactionally persisted in an outbox as part of the data-tier processing).
+        /// </summary>
+        [JsonProperty("eventOutbox", DefaultValueHandling = DefaultValueHandling.Ignore)]
+        [PropertySchema("Events", Title = "The the data-tier event outbox persistence technology (where the events will be transactionally persisted in an outbox as part of the data-tier processing).", IsImportant = true, Options = new string[] { "None", "Database" },
+            Description = "Defaults to `Entity.EventOutbox` configuration property (inherits) where not specified and `EventPublish` is `Data`; otherwise, `None`. A value of `Database` will result in the `DatabaseEventOutboxInvoker` being used to orchestrate.")]
+        public string? EventOutbox { get; set; }
 
         /// <summary>
         /// Gets or sets the URI event source.
@@ -677,6 +693,31 @@ operations: [
         public string? EventFormatKey { get; private set; }
 
         /// <summary>
+        /// Indicates whether the manual (OnImplementation) can be implemened as shorthand - minimal amount of code.
+        /// </summary>
+        public bool IsManualShorthand => AutoImplement == "None";
+
+        /// <summary>
+        /// Indicates whether any of the operations will raise an event within the Data-layer.
+        /// </summary>
+        public bool HasDataEvents => EventPublish == "Data";
+
+        /// <summary>
+        /// Gets or sets the DataInvoker code.
+        /// </summary>
+        public string DataInvoker { get; set; } = "DataInvoker.Current";
+
+        /// <summary>
+        /// Gets or sets the DataInvokerArgs code.
+        /// </summary>
+        public string DataInvokerArgs { get; set; } = "BusinessInvokerArgs";
+
+        /// <summary>
+        /// Indicates whether a send as well as a publish should occur.
+        /// </summary>
+        public bool DataEventSend { get; set; } = true;
+
+        /// <summary>
         /// <inheritdoc/>
         /// </summary>
         protected override void Prepare()
@@ -816,6 +857,7 @@ operations: [
 
             EventSource = DefaultWhereNull(EventSource, () => Parent!.EventSource);
             EventPublish = DefaultWhereNull(EventPublish, () => new string[] { "Create", "Update", "Delete" }.Contains(Type) ? Parent!.EventPublish : "None");
+            EventOutbox = DefaultWhereNull(EventOutbox, () => EventPublish != "None" ? Parent!.EventOutbox : "None");
 
             EventFormatKey = Type switch
             {
@@ -937,7 +979,7 @@ operations: [
         /// </summary>
         private void PrepareEvents()
         {
-            if (string.IsNullOrEmpty(EventSubject) || !CompareNullOrValue(Parent!.EventPublish, "None"))
+            if (string.IsNullOrEmpty(EventSubject) || CompareNullOrValue(EventPublish, "None"))
                 return;
 
             foreach (var @event in EventSubject!.Split(";", StringSplitOptions.RemoveEmptyEntries))
@@ -987,30 +1029,60 @@ operations: [
         private void PrepareData()
         {
             DataArgs = new ParameterConfig { Name = "<internal>", PrivateName = "__dataArgs" };
-            switch (AutoImplement)
+            switch (AutoImplement != "None" ? AutoImplement : Parent!.AutoImplement)
             {
                 case "Database":
                     DataArgs.Name = "_db";
                     DataArgs.Type = "IDatabaseArgs";
+
+                    if (EventOutbox != "None" && EventOutbox != "Database")
+                        throw new CodeGenException(this, nameof(EventOutbox), $"An Operation.AutoImplement (or Entity.AutoImplement) of 'Database' is at odds with the EventOutbox persistence of '{EventOutbox}'.");
+
                     break;
 
                 case "EntityFramework":
                     DataArgs.Name = "_ef";
                     DataArgs.Type = "IEfDbArgs";
+
+                    if (EventOutbox != "None" && EventOutbox != "Database")
+                        throw new CodeGenException(this, nameof(EventOutbox), $"An Operation.AutoImplement (or Entity.AutoImplement) of 'EntityFramework' is at odds with the EventOutbox persistence of '{EventOutbox}'.");
+
                     break;
 
                 case "Cosmos":
                     DataArgs.Name = "_cosmos";
                     DataArgs.Type = "ICosmosDbArgs";
+
+                    if (EventOutbox != "None" && EventOutbox != "Cosmos")
+                        throw new CodeGenException(this, nameof(EventOutbox), $"An Operation.AutoImplement (or Entity.AutoImplement) of 'Cosmos' is at odds with the EventOutbox persistence of '{EventOutbox}'.");
+
                     break;
 
                 case "OData":
                     DataArgs.Name = "_odata";
                     DataArgs.Type = "IODataArgs";
+
+                    if (EventOutbox != "None" && EventOutbox != "OData")
+                        throw new CodeGenException(this, nameof(EventOutbox), $"An Operation.AutoImplement (or Entity.AutoImplement) of 'OData' is at odds with the EventOutbox persistence of '{EventOutbox}'.");
+
+                    break;
+
+                default:
+                    if (EventPublish == "Data")
+                        throw new CodeGenException(this, nameof(EventPublish), "Unable to determine the EventOutbox 'Data' repository as both the Operation.AutoImplement and Entity.AutoImplement are set to 'None'; at least one of these must be set.");
+
                     break;
             }
 
             DataArgs.Prepare(Root!, this);
+
+            if (EventOutbox != "None")
+            {
+                DataInvoker = $"{(DataArgs.Name == "<internal>" ? "_db" : DataArgs.Name)}.EventOutboxInvoker";
+                DataInvokerArgs = "DatabaseEventOutboxInvokerArgs";
+                DataTransaction = false;
+                DataEventSend = false;
+            }
         }
 
         /// <summary>